# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Get All room",
                "operationId": "ea5014f568762b3074db88efca04c9d8",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "allowEmptyValue": true,
                        "example": "L222"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room retrieve succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "title": "data",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Room"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                }
            },
            "post": {
                "tags": [
                    "rooms"
                ],
                "summary": "Add new Room",
                "operationId": "ac14ad433d421c616b3c5cb81a3c4b3d",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Room data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "L222"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sensor created succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rooms/{roomId}": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Update Room",
                "operationId": "16cdcbaded391575376c6800993d3283",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "Id Of selected room",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room updated succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "rooms"
                ],
                "summary": "Update Room",
                "operationId": "ab7dfaa7c40d1610e628726c5e360c42",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "Id Of selected room",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Room data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "L222"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated succesfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "rooms"
                ],
                "summary": "Update Room",
                "operationId": "7af65cd06d7d3da3b4fe030817cafacf",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {access-token}",
                        "required": true,
                        "schema": {
                            "type": "bearerAuth"
                        }
                    },
                    {
                        "name": "roomId",
                        "in": "path",
                        "description": "Id Of selected room",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room deleted succesfully"
                    },
                    "403": {
                        "description": "Not Allowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Room": {
                "title": "schema of room",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}